requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
ics>=0.7
python-dateutil>=2.8.0
Flask>=2.0.0
playwright>=1.40.0
